fix modal not closing after form submit and refactor code
