--- src/crm/agents/agents.css
@@ -1,4 +1,0 @@
-.agents-container {
-  width: 90%;
-  margin-top: 3%;
-}
