--- src/crm/opportunities/opportunities.css
@@ -1,4 +1,0 @@
-.opportunities-container {
-  width: 90%;
-  margin-top: 3%;
-}
