+++ src/crm/opportunity-modal/opportunity-modal.html
@@ -48,14 +48,14 @@
                   <div class="form-group">
                     <label for="message-text" class="col-form-label">Stage</label>
                     <select class="form-control" value.bind="stage">
-                      <option value="New">New</option>
-                      <option value="Proposition">Proposition</option>
-                      <option value="Won">Won</option>
+                      <option value="new">New</option>
+                      <option value="proposition">Proposition</option>
+                      <option value="won">Won</option>
                     </select>
                   </div>
                   <div class="modal-footer">
                     <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                    <button type="submit" class="btn btn-primary">Save changes</button>
+                    <button type="button" class="btn btn-primary" click.delegate="newOpportunity(contact)" data-dismiss="modal">Save changes</button>
                   </div>
                 </form>
               </div>
