+++ src/crm/pipeline/pipeline.html
@@ -1,5 +1,5 @@
 <template>
-  <require from="/Users/siret/code/tarkvaratehnika/ofbiz-ui/src/crm/opportunity-modal/opportunity-modal.html"></require>
+  <require from="crm/opportunity-modal/opportunity-modal"></require>
   <div class="container mt-4">
     <opportunity-modal></opportunity-modal>
