+++ src/crm/service/opportunities-service.js
@@ -2,20 +2,19 @@ import { HttpClient, json } from 'aurelia-fetch-client';
 import { inject } from 'aurelia-framework';
 
 @inject(HttpClient)
-export class PipelineService {
+export class OpportunitiesService {
   constructor() {
     this.client = new HttpClient;
     const baseUrl = '/api/';
 
     this.client.configure(config => {
-      config
-        .withBaseUrl(baseUrl)
-    });
+      config.withBaseUrl(baseUrl);
+    })
   }
 
-  getNewOpportunities() {
+  getOpportunities() {
     return this.client
-      .fetch('opportunity/1/new')
+      .fetch('opportunities')
       .then(response => response.json())
       .catch(reason => {
         console.error(reason);
@@ -23,19 +22,9 @@ export class PipelineService {
       });
   }
 
-  getPropositionOpportunities() {
+  getOpportunity(id) {
     return this.client
-      .fetch('opportunity/1/proposition')
-      .then(response => response.json())
-      .catch(reason => {
-        console.error(reason);
-        return [];
-      });
-  }
-
-  getWonOpportunities() {
-    return this.client
-      .fetch('opportunity/1/won')
+      .fetch('opportunities')
       .then(response => response.json())
       .catch(reason => {
         console.error(reason);
