+++ src/crm/service/pipeline-service.js
@@ -1,20 +1,21 @@
-import { HttpClient } from 'aurelia-fetch-client';
+import { HttpClient, json } from 'aurelia-fetch-client';
 import { inject } from 'aurelia-framework';
 
 @inject(HttpClient)
-export class OpportunitiesService {
+export class PipelineService {
   constructor() {
     this.client = new HttpClient;
     const baseUrl = '/api/';
 
     this.client.configure(config => {
-      config.withBaseUrl(baseUrl);
-    })
+      config
+        .withBaseUrl(baseUrl)
+    });
   }
 
-  getOpportunities() {
+  getNewOpportunities() {
     return this.client
-      .fetch('opportunities')
+      .fetch('opportunity/1/new')
       .then(response => response.json())
       .catch(reason => {
         console.error(reason);
@@ -22,9 +23,19 @@ export class OpportunitiesService {
       });
   }
 
-  getOpportunity(id) {
+  getPropositionOpportunities() {
     return this.client
-      .fetch('opportunities')
+      .fetch('opportunity/1/proposition')
+      .then(response => response.json())
+      .catch(reason => {
+        console.error(reason);
+        return [];
+      });
+  }
+
+  getWonOpportunities() {
+    return this.client
+      .fetch('opportunity/1/won')
       .then(response => response.json())
       .catch(reason => {
         console.error(reason);
